updating

+++ management/server/core/security-manager/security-manager-impl/src/main/java/io/subutai/core/security/impl/crypto/KeyManagerImpl.java
@@ -640,9 +640,12 @@ public class KeyManagerImpl implements KeyManager
             {
                 String fingerprint = getFingerprint( identityId );
 
-                securityDataService.removeKeyData( identityId );
-                securityDataService.removeSecretKeyData( identityId );
-                securityDataService.removeKeyAllTrustData( fingerprint );
+                if ( fingerprint != null )
+                {
+                    securityDataService.removeKeyData( identityId );
+                    securityDataService.removeSecretKeyData( identityId );
+                    securityDataService.removeKeyAllTrustData( fingerprint );
+                }
             }
         }
         catch ( Exception ex )
